input_driver.state 返値を扱う部分の32bit拡張

+++ input/input_driver.h
@@ -904,7 +904,7 @@ void input_config_get_bind_string_joykey(
       char *buf, const char *prefix,
       const struct retro_keybind *bind, size_t size);
 
-int16_t input_state_internal(unsigned port, unsigned device,
+int32_t input_state_internal(unsigned port, unsigned device,
       unsigned idx, unsigned id);
 
 bool input_key_pressed(int key, bool keyboard_pressed);
@@ -1020,7 +1020,7 @@ void input_driver_poll(void);
  * Returns: Non-zero if the given key (identified by @id)
  * was pressed by the user (assigned to @port).
  **/
-int16_t input_driver_state_wrapper(unsigned port, unsigned device,
+int32_t input_driver_state_wrapper(unsigned port, unsigned device,
       unsigned idx, unsigned id);
 
 /**
